pyproject.toml
[project]
name = "language-processor"
version = "0.1.0"
description = "Add your description here"
authors = [
    { name = "Satont", email = "satontworldwide@gmail.com" }
]
dependencies = [
    "googletrans>=4.0.2",
    "uvicorn>=0.34.0",
    "fastapi>=0.115.12",
    "fasttext>=0.9.3",
    "numpy<2",
    "pybind11>=2.13.6",
    "scipy>=1.15.2",
    "iso639-lang>=2.6.0",
]
readme = "README.md"
requires-python = ">= 3.8"

[tool.rye.scripts]
server = "uvicorn app.server:app"
dev-server = "uvicorn app.server:app --reload"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.rye]
managed = true
virtual = true
universal = true
dev-dependencies = []
excluded-dependencies = []
